split http request funcs to separate file
